Move specific scrapers under global gem namespace
