+++ lib/quiz_scraper.rb
@@ -3,7 +3,7 @@ require "quiz_scraper/scraper_loader"
 require "quiz_scraper/scraper"
 require "quiz_scraper/request_handler"
 require "quiz_scraper/request_processor"
-require "clients/pub_quizzer"
+require "quiz_scraper/scrapers/pub_quizzer"
 
 module QuizScraper
 end
